refactor: `TableGroup` - `OrderTable` 양방향 의존성 제거
