+++ src/main/java/kitchenpos/table/domain/TableGroup.java
@@ -1,16 +1,11 @@
 package kitchenpos.table.domain;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
-import kitchenpos.table.exception.OrderTableCountNotEnoughException;
 
 @Table(name = "table_group")
 @Entity
@@ -22,8 +17,6 @@ public class TableGroup {
 
     private LocalDateTime createdDate;
 
-    @OneToMany(mappedBy = "tableGroup", cascade = CascadeType.ALL)
-    private List<OrderTable> orderTables = new ArrayList<>();
 
     protected TableGroup() {
     }
@@ -32,36 +25,11 @@ public class TableGroup {
         this.createdDate = createdDate;
     }
 
-    public void groupOrderTables(List<OrderTable> orderTables) {
-        for (OrderTable orderTable : orderTables) {
-            orderTable.validateIsEmpty();
-            orderTable.validateTableGroupNotExists();
-        }
-        this.orderTables = orderTables;
-        validateOrderTableCount();
-    }
-
-    private void validateOrderTableCount() {
-        if (orderTables.size() < 2) {
-            throw new OrderTableCountNotEnoughException();
-        }
-    }
-
-    public void unGroupOrderTables() {
-        for (OrderTable orderTable : orderTables) {
-            orderTable.changeTableGroup(null);
-        }
-    }
-
     public Long getId() {
         return id;
     }
 
     public LocalDateTime getCreatedDate() {
         return createdDate;
-    }
-
-    public List<OrderTable> getOrderTables() {
-        return orderTables;
     }
 }
