+++ src/main/java/kitchenpos/table/dto/response/TableGroupResponse.java
@@ -3,6 +3,7 @@ package kitchenpos.table.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
 
 public class TableGroupResponse {
@@ -17,8 +18,8 @@ public class TableGroupResponse {
         this.orderTables = orderTables;
     }
 
-    public static TableGroupResponse from(TableGroup tableGroup) {
-        List<OrderTableResponse> orderTableResponses = tableGroup.getOrderTables().stream()
+    public static TableGroupResponse from(TableGroup tableGroup, List<OrderTable> orderTables) {
+        List<OrderTableResponse> orderTableResponses = orderTables.stream()
                 .map(OrderTableResponse::from)
                 .collect(Collectors.toList());
