+++ src/main/java/kitchenpos/table/repository/OrderTableRepository.java
@@ -1,8 +1,11 @@
 package kitchenpos.table.repository;
 
+import java.util.List;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
 
+    List<OrderTable> findAllByTableGroup(TableGroup tableGroup);
 }
